+++ elastic-apm-agent/pom.xml
@@ -27,6 +27,11 @@
             <artifactId>apm-agent-core</artifactId>
             <version>${project.version}</version>
         </dependency>
+        <dependency>
+            <groupId>${project.groupId}</groupId>
+            <artifactId>apm-indy-bootstrap-module</artifactId>
+            <version>${project.version}</version>
+        </dependency>
 
         <!--Agent plugins, sorted alphabetically -->
         <dependency>
